improve spring-security-xsuaa-usage
